Revert "LPD-21704 Move package"
This reverts commit 119e627827545485de900a564e227d94bf949a65.
